+++ changelog.md
@@ -16,6 +16,7 @@ Saves from 8.x are not compatible with 9.0.0.
 * **[Mission Generation]** Configured target and initial points for F-15E S4+.
 * **[Mission Generation]** Added a package kneeboard page that shows the radio frequencies, tasks, and laser codes for each member of your package.
 * **[Mission Generation]** Added option to generate AI flights with unlimited fuel (enabled by default).
+* **[Mission Generation]** Added option to scale down unit count in mission generator.
 * **[Modding]** Factions can now specify the ship type to be used for cargo shipping. The Handy Wind will be used by default, but WW2 factions can pick something more appropriate.
 * **[Modding]** Unit variants can now set a display name separate from their ID.
 * **[Modding]** Updated Community A-4E-C mod version support to 2.2.0 release.